refactor(ChessGame): Game 메서드 및 예외 정리
