+++ src/main/java/chess/domain/ChessGame.java
@@ -6,34 +6,51 @@ import chess.domain.piece.Color;
 import chess.domain.piece.Piece;
 import chess.domain.piece.PieceType;
 import chess.domain.position.Position;
-import java.util.List;
 
 public class ChessGame implements Action {
     
-    
-    public static final String GAME_CANNOT_EXECUTE_MESSAGE = "실행할 수 없는 명령입니다.";
     public static final String GAME_HAS_NOT_STARTED = "게임이 시작되지 않았습니다.";
+    public static final String GAME_ALREADY_HAS_STARTED = "게임이 이미 시작되었습니다.";
+    private static final String COMMAND_ERROR_PREFIX = "[GAME ERROR] ";
     private final Board board;
     private Color turn;
     private GameStatus status;
     
     public ChessGame() {
-        this.status = GameStatus.START;
+        this.status = GameStatus.READY;
         this.board = Board.create();
         this.turn = Color.WHITE;
     }
     
-    
-    public void move(final List<String> arguments) {
-        if (this.status != GameStatus.MOVE) {
-            throw new IllegalStateException(GAME_HAS_NOT_STARTED);
+    @Override
+    public void start() {
+        if (this.isNotReady()) {
+            throw new IllegalArgumentException(COMMAND_ERROR_PREFIX + GAME_ALREADY_HAS_STARTED);
         }
-        Position source = Position.from(arguments.get(0));
-        Position destination = Position.from(arguments.get(1));
-        
-        this.checkPieceMove(source, destination);
-        this.board.replace(source, destination);
+        this.board.initialize();
+        this.status = GameStatus.RUN;
+    }
+    
+    @Override
+    public void move(final Position from, final Position to) {
+        if (this.isNotRunning()) {
+            throw new IllegalArgumentException(COMMAND_ERROR_PREFIX + GAME_HAS_NOT_STARTED);
+        }
+        this.checkPieceMove(from, to);
+        this.board.replace(from, to);
         this.turn = Color.reverse(this.turn);
+    }
+    
+    @Override
+    public void end() {
+        if (this.isNotRunning()) {
+            throw new IllegalArgumentException(COMMAND_ERROR_PREFIX + GAME_HAS_NOT_STARTED);
+        }
+        this.status = GameStatus.END;
+    }
+    
+    public boolean isNotRunning() {
+        return this.status != GameStatus.RUN;
     }
     
     private void checkPieceMove(final Position source, final Position destination) {
@@ -56,37 +73,13 @@ public class ChessGame implements Action {
         }
     }
     
-    @Override
-    public void start() {
-        if (this.status != GameStatus.START) {
-            throw new IllegalStateException(GAME_CANNOT_EXECUTE_MESSAGE);
-        }
-        this.board.initialize();
-        this.status = GameStatus.MOVE;
-    }
-    
-    @Override
-    public void move(final Position from, final Position to) {
-        if (this.status != GameStatus.MOVE) {
-            throw new IllegalStateException(GAME_HAS_NOT_STARTED);
-        }
-        this.checkPieceMove(from, to);
-        this.board.replace(from, to);
-        this.turn = Color.reverse(this.turn);
-    }
-    
-    @Override
-    public void end() {
-        if (this.status != GameStatus.MOVE) {
-            throw new IllegalStateException(GAME_HAS_NOT_STARTED);
-        }
-        this.status = GameStatus.END;
+    public boolean isNotReady() {
+        return this.status != GameStatus.READY;
     }
     
     public boolean isNotEnd() {
         return this.status != GameStatus.END;
     }
-    
     
     public Board getBoard() {
         return this.board;
