+++ src/main/java/chess/view/OutputView.java
@@ -4,7 +4,6 @@ import chess.controller.BoardDTO;
 
 public class OutputView {
     
-    public static final String ERROR_PREFIX = "[ERROR] ";
     private static final String GAME_START_MESSAGE = "> 체스 게임을 시작합니다.\n"
             + "> 게임 시작 : start\n"
             + "> 게임 종료 : end\n"
@@ -22,7 +21,7 @@ public class OutputView {
     }
     
     public void printError(final String message) {
-        System.out.println(ERROR_PREFIX + message);
+        System.out.println(message);
         System.out.println();
     }
 }
